refactor(home): onboarding page
